+++ src/utils.rs
@@ -1,3 +1,6 @@
+use color_eyre::eyre::{WrapErr as _, eyre};
+use rustls_pki_types::CertificateDer;
+use rustls_pki_types::pem::PemObject as _;
 use std::borrow::Cow;
 use std::fmt::Write;
 use std::sync::OnceLock;
@@ -104,6 +107,47 @@ fn rewrite_to_positional(sql: &str) -> Cow<'_, str> {
     Cow::Owned(result)
 }
 
+/// Loads IACA root certificates from the given file paths.
+///
+/// Each path may point to either a DER-encoded certificate or a PEM file containing
+/// one or more certificates.  Format is detected by checking whether the file starts
+/// with the ASCII bytes `-----BEGIN`.
+///
+/// DER files are accepted without structural validation; a malformed DER file
+/// produces an `InvalidCertificateChain` error at the first credential
+/// verification, not at startup.  PEM files are validated only insofar as they
+/// must contain at least one parseable `CERTIFICATE` block.
+///
+/// # Errors
+///
+/// Returns an error if any path cannot be read, if a PEM file is malformed, or if a
+/// PEM file contains no `CERTIFICATE` blocks.
+pub(crate) fn load_iaca_roots(paths: &[String]) -> color_eyre::Result<Vec<Vec<u8>>> {
+    let mut roots = Vec::new();
+    for path in paths {
+        let bytes = std::fs::read(path)
+            .wrap_err_with(|| format!("failed to read IACA root file '{path}'"))?;
+
+        if bytes.starts_with(b"-----BEGIN") {
+            let mut count = 0usize;
+            for cert in CertificateDer::pem_slice_iter(&bytes) {
+                let cert =
+                    cert.wrap_err_with(|| format!("malformed PEM in IACA root file '{path}'"))?;
+                roots.push(cert.as_ref().to_vec());
+                count += 1;
+            }
+            if count == 0 {
+                return Err(eyre!(
+                    "IACA root file '{path}' is PEM-formatted but contains no certificates"
+                ));
+            }
+        } else {
+            roots.push(bytes);
+        }
+    }
+    Ok(roots)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -115,5 +159,71 @@ mod tests {
             rewrite_to_positional(sql).as_ref(),
             "SELECT * FROM credentials WHERE id = ? AND tenant_id = ?"
         );
+    }
+
+    // --- load_iaca_roots tests ---
+
+    fn self_signed_cert() -> (Vec<u8>, String) {
+        let params =
+            rcgen::CertificateParams::new(vec!["test.local".to_string()]).expect("cert params");
+        let key = rcgen::KeyPair::generate().expect("key generation");
+        let cert = params.self_signed(&key).expect("self-signed cert");
+        (cert.der().to_vec(), cert.pem())
+    }
+
+    #[test]
+    fn loads_der_and_pem_to_same_bytes() {
+        use cloud_wallet_openid4vc::formats::mdoc::{IacaTrustStore, StaticTrustStore};
+        use std::io::Write as _;
+
+        let (der, pem) = self_signed_cert();
+        for (label, content) in [("DER", der.clone()), ("PEM", pem.into_bytes())] {
+            let mut f = tempfile::NamedTempFile::new().unwrap();
+            f.write_all(&content).unwrap();
+            let roots = load_iaca_roots(&[f.path().to_string_lossy().into_owned()]).unwrap();
+            assert_eq!(
+                roots,
+                vec![der.clone()],
+                "{label} file should decode to the same DER"
+            );
+            let store = StaticTrustStore::new(roots);
+            assert_eq!(
+                store.trusted_roots(),
+                std::slice::from_ref(&der),
+                "{label} roots must reach the store"
+            );
+        }
+    }
+
+    #[test]
+    fn pem_bundle_loads_all_certs() {
+        use std::io::Write as _;
+
+        let (der1, pem1) = self_signed_cert();
+        let (der2, pem2) = self_signed_cert();
+        let mut f = tempfile::NamedTempFile::new().unwrap();
+        f.write_all(format!("{pem1}{pem2}").as_bytes()).unwrap();
+        let roots = load_iaca_roots(&[f.path().to_string_lossy().into_owned()]).unwrap();
+        assert_eq!(roots.len(), 2);
+        assert!(roots.contains(&der1) && roots.contains(&der2));
+    }
+
+    #[test]
+    fn errors_on_bad_input() {
+        use std::io::Write as _;
+
+        assert!(load_iaca_roots(&["/nonexistent/path/root.pem".into()]).is_err());
+
+        let mut f1 = tempfile::NamedTempFile::new().unwrap();
+        f1.write_all(
+            b"-----BEGIN CERTIFICATE-----\nnot-valid-base64!!!\n-----END CERTIFICATE-----\n",
+        )
+        .unwrap();
+        assert!(load_iaca_roots(&[f1.path().to_string_lossy().into_owned()]).is_err());
+
+        let mut f2 = tempfile::NamedTempFile::new().unwrap();
+        f2.write_all(b"-----BEGIN PRIVATE KEY-----\nMIIE\n-----END PRIVATE KEY-----\n")
+            .unwrap();
+        assert!(load_iaca_roots(&[f2.path().to_string_lossy().into_owned()]).is_err());
     }
 }
